refactor: changed folder name from "wagmiDataSource" to rpc"
