+++ src/stores/modules/SnapshotERC20Guild/fetchers/rpc/useTotalLocked.ts
