+++ src/stores/modules/SnapshotRepGuild/fetchers/rpc/useTotalLocked.ts
